+++ src/main/java/fr/inra/maiage/bibliome/util/marshall/WriteCache.java
@@ -18,6 +18,6 @@ package fr.inra.maiage.bibliome.util.marshall;
 
 public interface WriteCache<T> {
 	boolean contains(T object);
-	int get(T object);
-	void put(T object, int position);
+	long get(T object);
+	void put(T object, long position);
 }
